requirements.txt
pandas==2.2.2
pyyaml==6.0.2
duckdb==1.1.3
pyarrow==17.0.0